refactor: Moves services to @/lib/services/ sub-dir
