+++ src/db/schema.ts
@@ -7,6 +7,7 @@ import {
   primaryKey,
   timestamp,
   boolean,
+  uuid,
 } from "drizzle-orm/pg-core";
 import { relations } from "drizzle-orm";
 
@@ -15,7 +16,7 @@ import { relations } from "drizzle-orm";
 // #########################################################
 
 export const users = pgTable("users", {
-  id: text("id").primaryKey(),
+  id: uuid("id").primaryKey().defaultRandom(),
   email: text("email").notNull().unique(),
   name: text("name"),
   createdAt: timestamp("created_at").notNull().defaultNow(),
