+++ src/lib/services/wishlist-service.ts
@@ -1,6 +1,6 @@
-import { wishlists, wishlistItems } from "../db/schema";
+import { wishlists, wishlistItems } from "@/db/schema";
 import { eq, and } from "drizzle-orm";
-import type { DrizzleDatabase } from "../db";
+import type { DrizzleDatabase } from "@/db";
 
 export interface CreateWishlistInput {
   name: string;
