+++ src/pages/api/user/bookmarked/[slug].ts
@@ -1,11 +1,17 @@
 import type { APIRoute } from "astro";
 import { db } from "@/db";
-import { isGiftBookmarked } from "@/lib/bookmark-service";
-import { getGiftIdBySlug } from "@/lib/gift-service";
-import { getCurrentUser } from "@/lib/user-service";
+import { isGiftBookmarked } from "@/lib/services/bookmark-service";
+import { getGiftIdBySlug } from "@/lib/services/gift-service";
+import { getCurrentUser } from "@/lib/services/user-service";
 
 export const prerender = false;
 
+/**
+ * Check if a gift is bookmarked by the current user
+ * @param params - The parameters from the URL
+ * @param request - The request object
+ * @returns A response object
+ */
 export const GET: APIRoute = async ({ params, request }) => {
   const user = await getCurrentUser(db, request);
   if (!user) {
