+++ src/pages/index.astro
@@ -1,11 +1,11 @@
 ---
 // src/pages/index.astro
-import Layout from "../layouts/Layout.astro";
-import HeroBanner from "../components/HeroBanner.astro";
-import GiftGallery from "../components/GiftGallery.astro";
-import { getGiftsByCategory } from "../lib/gift-service";
-import { GiftCategory } from "../content/types";
-import { db } from "../db";
+import Layout from "@/layouts/Layout.astro";
+import HeroBanner from "@/components/HeroBanner.astro";
+import GiftGallery from "@/components/GiftGallery.astro";
+import { getGiftsByCategory } from "@/lib/services/gift-service";
+import { GiftCategory } from "@/content/types";
+import { db } from "@/db";
 
 const categories = [
   { id: GiftCategory.enum.subscription, title: "Subscriptions" },
